Fix ID Generation, add gifv support.
idList will hold all id's, so we are sure we have a unique ID. If an attempt to create an id begins, and the ID already exists, retry.

A first attempt to add gifv support. The original developer appended .jpg to any image without an extension, but did not update his regular expression for Gifv, so links would be .gifv.jpg instead of only .gifv.